requirements.txt
pandas
numpy
python-dotenv
openai
pyarrow
openpyxl
datasets
